Move types from Cardano.CLI.Legacy.Commands to Cardano.CLI.Types.Common
