+++ cardano-cli/src/Cardano/CLI/Byron/Key.hs
@@ -18,7 +18,6 @@ where
 import           Cardano.Api.Byron
 
 import qualified Cardano.Chain.Common as Common
-import           Cardano.CLI.Legacy.Commands (ByronKeyFormat (..))
 import           Cardano.CLI.Types.Common
 import qualified Cardano.Crypto.Signing as Crypto
