+++ cardano-cli/src/Cardano/CLI/Byron/Run.hs
@@ -20,7 +20,6 @@ import           Cardano.CLI.Byron.Tx
 import           Cardano.CLI.Byron.UpdateProposal
 import           Cardano.CLI.Byron.Vote
 import           Cardano.CLI.Helpers
-import           Cardano.CLI.Legacy.Commands (ByronKeyFormat (..))
 import           Cardano.CLI.Types.Common
 import qualified Cardano.Crypto.Hashing as Crypto
 import qualified Cardano.Crypto.Signing as Crypto
