+++ cardano-cli/src/Cardano/CLI/Legacy/Commands.hs
@@ -470,105 +470,3 @@ renderLegacyGenesisCmds cmd =
     GenesisTxIn {} -> "genesis initial-txin"
     GenesisAddr {} -> "genesis initial-addr"
     GenesisHashFile {} -> "genesis hash"
-
---
--- Shelley CLI flag/option data types
---
-
-newtype ProtocolParamsFile
-  = ProtocolParamsFile FilePath
-  deriving (Show, Eq)
-
-newtype TxInCount
-  = TxInCount Int
-  deriving Show
-
-newtype TxOutCount
-  = TxOutCount Int
-  deriving Show
-
-newtype TxShelleyWitnessCount
-  = TxShelleyWitnessCount Int
-  deriving Show
-
-newtype TxByronWitnessCount
-  = TxByronWitnessCount Int
-  deriving Show
-
-newtype BlockId
-  = BlockId String -- Probably not a String
-  deriving Show
-
-newtype GenesisKeyFile
-  = GenesisKeyFile FilePath
-  deriving Show
-
-data MetadataFile = MetadataFileJSON (File () In)
-                  | MetadataFileCBOR (File () In)
-
-  deriving Show
-
-type StakePoolMetadataFile = File StakePoolMetadata
-
-newtype GenesisDir
-  = GenesisDir FilePath
-  deriving Show
-
--- | Either a verification or signing key, used for conversions and other
--- commands that make sense for both.
---
-data SomeKeyFile direction
-  = AVerificationKeyFile (VerificationKeyFile direction)
-  | ASigningKeyFile (SigningKeyFile direction)
-  deriving Show
-
-data AddressKeyType
-  = AddressKeyShelley
-  | AddressKeyShelleyExtended
-  | AddressKeyByron
-  deriving Show
-
-data ByronKeyType
-  = ByronPaymentKey  ByronKeyFormat
-  | ByronGenesisKey  ByronKeyFormat
-  | ByronDelegateKey ByronKeyFormat
-  deriving Show
-
-data ByronKeyFormat = NonLegacyByronKeyFormat
-                    | LegacyByronKeyFormat
-  deriving Show
-
--- | The type of @cardano-address@ key.
-data CardanoAddressKeyType
-  = CardanoAddressShelleyPaymentKey
-  | CardanoAddressShelleyStakeKey
-  | CardanoAddressIcarusPaymentKey
-  | CardanoAddressByronPaymentKey
-  deriving Show
-
-type OpCertCounterFile = File OpCertCounter
-
-newtype PrivKeyFile
-  = PrivKeyFile FilePath
-  deriving Show
-
-newtype WitnessFile
-  = WitnessFile FilePath
-  deriving Show
-
--- | A raw verification key given in Base64, and decoded into a ByteString.
-newtype VerificationKeyBase64
-  = VerificationKeyBase64 String
-  deriving Show
-
--- | Data required to construct a witness.
-data WitnessSigningData
-  = KeyWitnessSigningData
-      !(SigningKeyFile In)
-      -- ^ Path to a file that should contain a signing key.
-      !(Maybe (Address ByronAddr))
-      -- ^ An optionally specified Byron address.
-      --
-      -- If specified, both the network ID and derivation path are extracted
-      -- from the address and used in the construction of the Byron witness.
-  deriving Show
